// src/api/models/tasks-extract-task-details-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TasksExtractTaskDetailsCreateRequest
 */
export interface TasksExtractTaskDetailsCreateRequest {
    /**
     * Natural language description of the task
     * @type {string}
     * @memberof TasksExtractTaskDetailsCreateRequest
     */
    'description': string;
}